bsp: simplify nrf53 specific implementation of rng bsp library
