+++ nrf5340dk-net.emProject
@@ -49,7 +49,7 @@
     linker_printf_fp_enabled="Float"
     linker_printf_width_precision_supported="Yes"
     linker_section_placement_file="$(ProjectDir)/../../nRF/Setup/nRF5340_xxAA_Network_flash_placement.xml"
-    macros="BuildTarget=nrf5340dk-net;Lh2ImplementationFile=lh2.c;PwmImplementationFile=pwm_nrf5340_net.c;RadioImplementationFile=radio.c;RngImplementationFile=rng.c;TdmaClientImplementationFile=tdma_client.c;TdmaServerImplementationFile=tdma_server.c;DeviceFamily=nRF;Target=nRF5340_xxAA_Network;Placement=Flash"
+    macros="BuildTarget=nrf5340dk-net;Lh2ImplementationFile=lh2.c;PwmImplementationFile=pwm_nrf5340_net.c;RadioImplementationFile=radio.c;TdmaClientImplementationFile=tdma_client.c;TdmaServerImplementationFile=tdma_server.c;DeviceFamily=nRF;Target=nRF5340_xxAA_Network;Placement=Flash"
     project_type="Executable"
     target_reset_script="Reset();"
     target_trace_initialize_script="EnableTrace(&quot;$(TraceInterfaceType)&quot;)" />
